fix: added useTooltip hook
