+++ src/hooks/index.ts
@@ -33,3 +33,4 @@ export * from './xcm/useXcmAssets';
 export * from './xcm/useXcmBridge';
 export * from './xcm/useXcmTokenTransfer';
 export * from './xcm/useXcmTokenDetails';
+export * from './ui/useTooltip';
